chore(tests): split rss.test.ts into rss + rss-ext and rename shared to edge-cases
- Extract /feeds/author/:author.xml and /feeds/lang/:lang.xml describe blocks
  from rss.test.ts (388 lines) into new rss-ext.test.ts to satisfy 300-line cap
  (rss.test.ts: 203 lines, rss-ext.test.ts: 258 lines)
- Rename shared.test.ts -> edge-cases.test.ts for clarity; content unchanged

Closes reviewer feedback on PR #371
